requirements.txt
telethon==1.17.5
python-dotenv==1.0.0
python-decouple==3.8